Updated
Looked over and added the functions pay_out() and pay_in() into class Account, but both are blank float functions. 

Don't know if addAccount or findAccount are working because the program does not compile. 

The issue with the findAccount if statement (infinity loop?)

Added the class Database from the prompt, but it doesn't recognize the const Account &a reference (should be on my end due to nameing my variables, but I haven't found the workaround yet)

This is the one I need the most help with.